Better RelationshipTypeEditor integration
